requirements.txt
matplotlib==3.3.4
motmetrics==1.1.3
torch==1.12.0
numba==0.54.1
numpy==1.20.0
numpyencoder==0.3.0
nuscenes_devkit==1.1.9
opencv_python==4.5.3.56
pyquaternion==0.9.9
scipy==1.6.2
Shapely==1.7.1
tqdm==4.59.0
